requirements.txt
onnx
numpy
gym
gym[atari]
atari-py
gym[box2d]
onnx2pytorch
seaborn
matplotlib
pyglet==1.3.2
-e git+https://github.com/ministry-of-silly-code/experiment_buddy.git#egg=experiment_buddy